Add a check target for formatting in flake
Ensures the version of ormolu used to check files is
the same as the one provided by the environment (and hence used to
format files).
Also update the flake lock: the files were formatted with ormolu 0.5 but
the nixpkgs version was providing ormolu 0.3.
Also use `cabal test all` in the CI rather than one test for each backend